test 1 - refactor
