+++ src/Domain/VolleyM.Domain.Contracts/Crosscutting/IEventHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace VolleyM.Domain.Framework.EventBus
+namespace VolleyM.Domain.Contracts.Crosscutting
 {
     public interface IEventHandler<T>
     {
